added wrap on smaller screens
fixed position of the header on smaller screens
